requirements.txt
Django==1.10.1
django-multiselectfield==0.1.4
django-bootstrap3-datetimepicker-2==2.4.2
django-crispy-forms==1.6.1
django-bootstrap-form==3.2.1